pyproject.toml
[project]
name = "credit-risk-decision-engine"
version = "0.1.0"
description = "Credit risk decisioning API (PD + decision + reason codes) with evaluation-as-tests."
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "pandas>=2.2",
  "numpy>=1.26",
  "scikit-learn>=1.5",
  "lightgbm>=4.5",
  "mlflow>=2.15",
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.8",
  "pandera>=0.20",
  "pyyaml>=6.0",
  "python-dotenv>=1.0",
  "shap>=0.46",
  "prometheus-client>=0.20",
  "httpx>=0.27",
  "matplotlib>=3.8"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "black>=24.4",
  "mypy>=1.10",
  "pre-commit>=3.7"
]


[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

